docker_images/crawler/spider/requirements.txt
pybloomfiltermmap3==0.5.5
requests==2.25.1
requests-oauthlib==1.3.1
lxml==4.9.1
pika==1.3.0
pika-stubs==0.1.3
retry==0.9.2
redis==2.10.6